requirements.txt
Flask
pillow
torch
torchvision
gunicorn
matplotlib